backend/requirements.txt
fastapi
uvicorn[standard]
playwright
bs4               # html parsing
python-dotenv
supabase>=2       # you'll use later for direct inserts
openai>=1.6.1     # for embeddings
langchain         # for summarization
langchain-openai  # langchain-openai integration
pgvector          # for vector operations 